+++ addons/dialogic/Nodes/DialogNode.gd
@@ -38,10 +38,10 @@ var custom_events = {}
 
 ### DATA
 var definitions = {}
-var characters
 
 # Parsing results
 var questions #for keeping track of the questions answered
+var anchors = {} # for listing the indexes of the anchors
 
 ### CURRENT STATE
 var current_timeline: String = ''
@@ -126,9 +126,7 @@ func _ready():
 	$TextBubble/RichTextLabel.connect('meta_hover_started', self, '_on_RichTextLabel_meta_hover_started')
 	$TextBubble/RichTextLabel.connect('meta_hover_ended', self, '_on_RichTextLabel_meta_hover_ended')
 	
-	# Getting the character information
-	characters = DialogicUtil.get_character_list()
-
+	
 	if Engine.is_editor_hint():
 		if preview:
 			get_parent().connect("resized", self, "resize_main")
@@ -414,6 +412,7 @@ func load_dialog():
 		dialog_script = parse_characters(dialog_script)
 	dialog_script = parse_text_lines(dialog_script)
 	dialog_script = parse_branches(dialog_script)
+	parse_anchors()
 	return dialog_script
 
 # adds name coloring to the dialog texts
@@ -609,6 +608,13 @@ func _insert_glossary_definitions(text: String):
 		)
 	return final_text;
 
+func parse_anchors():
+	anchors = {}
+	var idx = 0
+	for event in dialog_script['events']:
+		if event['event_id'] == 'dialogic_015':
+			anchors[event['id']] = idx
+		idx += 1
 
 ## -----------------------------------------------------------------------------
 ## 					MAIN GAME-LOGIC 
@@ -883,6 +889,15 @@ func event_handler(event: Dictionary):
 			if event.get('set_random', false):
 				value = str(randi()%int(event.get("random_upper_limit", 100)-event.get('random_lower_limit', 0))+event.get('random_lower_limit', 0))
 			Dialogic.set_variable_from_id(event['definition'], value, operation)
+			_load_next_event()
+		# Anchor event
+		'dialogic_015':
+			emit_signal("event_start", "anchor", event)
+			_load_next_event()
+		# GoTo event
+		'dialogic_016':
+			emit_signal("event_start", "goto", event)
+			dialog_index = anchors[event.get('anchor_id')]
 			_load_next_event()
 		
 		
@@ -1449,6 +1464,7 @@ func dprint(string, arg1='', arg2='', arg3='', arg4='' ):
 
 # helper that allows to get a character by file
 func get_character(character_id):
+	var characters = DialogicUtil.get_character_list()
 	for c in characters:
 		if c['file'] == character_id:
 			return c
@@ -1491,8 +1507,9 @@ func _compare_definitions(def_value: String, event_value: String, condition: Str
 ## -----------------------------------------------------------------------------
 ## 					DIALOG FADING
 ## -----------------------------------------------------------------------------
-# ähm... no clue. What is this @emilio?
-# TODO wtf is this name?
+# Since Dialogic has some placeholder text in case something goes wrong
+# This will reset the text, reset any modulation it might have, and
+# set the variables that handle the fade in to the start position
 func _hide_dialog():
 	$TextBubble.update_text('') # Clearing the text
 	$TextBubble.modulate = Color(1,1,1,0)