+++ addons/dialogic/Documentation/Content/Changelog.md
@@ -3,6 +3,8 @@
 ## v1.4
 - Wait seconds event can now be set to be skipped with the user's action [[SimonLammer](https://github.com/SimonLammer)]
 - Added: `Autofocus choices` in the settings [[Jowan-Spooner](https://github.com/Jowan-Spooner)]
+- New events:
+  - `Label Event` and `Go to Event`. This will help you creating an anchor position to go back to.
 
 ## v1.3 - Save me some time
 #### General Editor Stuff